+++ settfex/__init__.py
@@ -35,7 +35,7 @@ Usage:
     >>> asyncio.run(main())
 """
 
-__version__ = "0.19.2"
+__version__ = "0.20.0"
 __author__ = "batt"
 __license__ = "MIT"